Resolve the cyclical value-dependency between tsickle.ts and transformer.ts by merging the two.
transformer mode now *is* tsickle, and the files are closely related in their imports, exports, and usage.
